// src/app/models/step-participants-model.ts
export class stepparticipantsmodel {
  Id: number;
  ParticipantId: number;
  Label: string;
  StepActionId: number;
  StepId: number;
  ParticipantName: string;
  WorkflowId: number;
  WorkflowName: string;
  StepActionIdList: string;
  ParameterCarrier: string;
  VisibilityFunction: string;
  Comments: string;
  ParticipantType: string;

}


export class MappingActionRoleViewModel{
  StepActionId: number;
  StepId: number;
  ProjectId: number;
  ColumnName: string;
  NewResponse: boolean;
  ParameterCarrier: string;
}